Add tests to go along with Const.
For all Const nodes (except children of Call nodes) all pairs of enclosing
paren are used/consumed/counted in the endings locations.
E.g. (((6))), ((1, ((True)))

However, when a Const node is a child of a Call node, n-1 pairs are used,
E.g. In the (extreme) example `range(((((2)))))`, only the parens in
`((((2))))` count towards the endings locations.

The only tests that currently fail for Const combine the above two examples,
`fun(2, (3, 2), (4))` and `print(False, ((True)))`...